move case mapping to it own class
